# docs/swagger.yaml
basePath: /
definitions:
  handlers.EndSessionRequest:
    properties:
      session_id:
        example: 123
        type: integer
    type: object
...
  handlers.StartSessionReq:
    properties:
      pc_number:
        type: integer
      tariff_id:
        type: integer
    type: object
  handlers.TariffBody:
    properties:
      duration:
        example: 10
        type: integer
      name:
        example: Free
        type: string
      price:
        example: 0
        type: number
      tariff_id:
        example: 123
        type: integer
    type: object
...
  models.Session:
    properties:
      end_time:
        type: string
      id:
        type: integer
      pc_number:
        type: integer
      start_time:
        type: string
      status:
        $ref: '#/definitions/models.SessionStatus'
      tariff_id:
        type: integer
      user_id:
        type: integer
    type: object
  models.SessionStatus:
    enum:
    - active
    - finished
    type: string
    x-enum-varnames:
    - Active
    - Finished
  models.Tariff:
    properties:
      created_at:
        type: string
      duration:
        type: integer
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample API with Swagger
  title: API Computer-club
  version: "1.0"
paths:
...
  /session/active:
    get:
      description: Возвращает список всех активных сессий (только для администраторов)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Session'
            type: array
        "403":
          description: 'ошибка доступа к текущему запросу. необходима роль пользователя:
            админ'
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение активных сессий
      tags:
      - sessions
  /session/end:
    post:
      consumes:
      - application/json
      description: Завершает активную сессию по её идентификатору
      parameters:
      - description: Данные для завершения сессии
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.EndSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "Session ended successfully'
          schema:
            type: string
        "400":
          description: Некорректный запрос (например, session_id == 0)
          schema:
            type: string
        "404":
          description: Сессия или компьютер не найдены
          schema:
            type: string
        "409":
          description: Сессия уже завершена или компьютер уже свободный
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера при закрытии сессии или транзакции
            базы данных или редис не удалил данные
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Завершение сессии
      tags:
      - sessions
  /session/start:
    post:
      consumes:
      - application/json
      description: Запускает новую сессию для пользователя на указанном ПК с выбранным
        тарифом
      parameters:
      - description: Данные для старта сессии
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.StartSessionReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Session'
        "400":
          description: ошибка JSON запроса
          schema:
            type: string
        "401":
          description: неправильный user_id в токене
          schema:
            type: string
        "404":
          description: Не найдено - user, computer или tariff или кошелек
          schema:
            type: string
        "409":
          description: Сессия уже активная или компьютер занят
          schema:
            type: string
        "500":
          description: внутренняя проблема поиска в базе данных или ошибка сохранения
            изменения транзакции базы данных или редис не создал данные
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Начало сессии
      tags:
      - sessions
  /tariff:
    get:
      description: Возвращает список всех доступных тарифов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Tariff'
            type: array
        "500":
          description: ошибка при поиске тарифов в базе данных
          schema:
            type: string
      summary: Получение списка тарифов
      tags:
      - tariffs
    post:
      consumes:
      - application/json
      description: Создает новый тариф с указанными данными (доступен для администратора)
      parameters:
      - description: Данные для тарифа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.TariffBody'
      produces:
      - application/json
      responses:
        "200":
          description: Тариф успешно создан
          schema:
            $ref: '#/definitions/models.Tariff'
        "400":
          description: Некорректные данные (пустое имя, отрицательный айди или цена
            или длительность)
          schema:
            type: string
        "403":
          description: 'Недостаточно прав: нужен пользователь с ролью админ'
          schema:
            type: string
        "409":
          description: Айди тарифа уже используется
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Создание нового тарифа
      tags:
      - tariffs
  /tariff/{id}:
    delete:
      description: удаление Тарифа из базы данных (доступно для админа)
      produces:
      - application/json
      responses:
        "200":
          description: ': "тариф успешно удален'
          schema:
            type: string
        "400":
          description: Некорректный ID тарифа
          schema:
            type: string
        "403":
          description: Ошибка доступа - недостаточно прав
          schema:
            type: string
        "404":
          description: Тариф не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: удаление Тарифа
      tags:
      - tariffs
    get:
      description: Возвращает информацию о тарифе по указанному ID
      parameters:
      - description: ID тарифа
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Tariff'
        "400":
          description: ошибка чтения тариф id
          schema:
            type: string
        "404":
          description: тариф не найден
          schema:
            type: string
        "500":
          description: внутренняя ошибка сервера
          schema:
            type: string
      summary: Получение тарифа по ID
      tags:
      - tariffs
...
swagger: "2.0"